refactor(web-components): Update GridItem and DragHandle styles and structure
- Renamed and reorganized styles for GridItem and DragHandle components, improving consistency and maintainability.
- Introduced new styles.scss files for both components, replacing the previous module styles.
- Updated GridItem and DragHandle components to use the new styles, enhancing visual presentation.
- Modified TrendChart theme to include additional properties for better customization.
- Removed unused DragHandleIcon styles to streamline the codebase.
