+++ packages/web-components/src/components/Grid/components/GridItem/GridItem.tsx
@@ -1,7 +1,7 @@
 import { clsx } from "clsx";
 import type { FunctionComponent } from "preact";
 import type { HTMLAttributes } from "preact/compat";
-import styles from "./GridItem.module.scss";
+import styles from "./styles.scss";
 import { DragHandle } from "./components";
 import { DEFAULT_DRAG_ENABLED_CLASSNAME } from "../../constants";
