--- packages/web-components/src/components/Grid/components/GridItem/components/DragHandle/components/DragHandleIcon.module.scss
@@ -1,3 +1,0 @@
-.drag-handle-icon {
-    flex: 0 0 auto;
-}
