+++ packages/web-components/src/components/Grid/components/GridItem/styles.scss
@@ -1,6 +1,6 @@
 .grid-item {
   display: grid;
-  grid-template-columns: minmax(0, 1fr) 28px;
+  grid-template-columns: minmax(0, 1fr) auto;
   width: 100%;
   height: 100%;
   box-sizing: border-box;
@@ -26,6 +26,6 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  width: 100%;
+  width: 28px;
   max-height: 28px;
 }
