+++ packages/web-components/src/components/TrendChart/theme.ts
@@ -1,12 +1,13 @@
 // Define the Nivo theme object using CSS variables
 export const nivoTheme = {
+  background: "var(--bg-base-primary)", // Chart background
   axis: {
     ticks: { // axis ticks (values on the axis)
       text: {
         fill: "var(--on-text-secondary)",
       },
     },
-    legend: { // legend text
+    legend: { // legend text (axis title)
       text: {
         fill: "var(--on-text-primary)",
       },
@@ -15,6 +16,17 @@ export const nivoTheme = {
       line: {
         stroke: "var(--on-border-muted)",
       },
+    },
+  },
+  legends: { // Symbol legends text (e.g., below the chart)
+    text: {
+      fill: "var(--on-text-secondary)",
+    },
+  },
+  tooltip: {
+    container: {
+      background: "var(--bg-base-modal)",
+      color: "var(--on-text-primary)",
     },
   },
 };
