+++ packages/web-dashboards/src/components/Dashboard/components/Widget/styles.scss
@@ -1,5 +1,4 @@
 .widget {
-  background: var(--bg-base-modal);
   border-radius: 8px;
   box-shadow: var(--shadow-small);
   overflow: hidden;
